\name{EmptyTokenFilter-class}
\docType{class}
\alias{EmptyTokenFilter-class}
\alias{show,EmptyTokenFilter-method}

\title{Class "EmptyTokenFilter"}

\description{
  A specialization of \code{\linkS4class{FunctionalTokenFilter}} that
  filters out all empty tokens.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{EmptyTokenFilter()}.
}

\section{Extends}{
Class \code{"\linkS4class{FunctionalTokenFilter}"}, directly.
Class \code{"\linkS4class{TokenFilter}"}, by class "FunctionalTokenFilter", distance 2.
Class \code{"\linkS4class{TokenTransform}"}, by class "FunctionalTokenFilter", distance 3.
}

\author{Anthony Fader, Gary King, Dan Pemstein, and Kevin Quinn}

\seealso{
  \code{\link{filterTokens}}
  \code{\linkS4class{FunctionalTokenFilter}} 
  \code{\linkS4class{TokenFilter}}
}


\keyword{classes}
