\name{predict.lda}
\alias{predict.lda}

\title{predict.lda}

\description{
  \code{predict.lda} is a generic function that uses a fitted
\code{\linkS4class{lda}} model to perform inference on novel text.
}

\usage{
  predict(object, newdata, var.max.iter = -1, var.convergence = 1e-06,
          em.max.iter = 100, em.convergence = 1e-04,
          estimate.alpha=TRUE)
}

\value{
  A documents (in \code{newdata}) by topics (in \code{object}) matrix
  of variational gamma parameters generated by applying the fitted
  model to the new data.
}

\arguments{
\item{object}{An \code{\linkS4class{lda}} fitted model.}
\item{newdata}{An object representing one or more documents to perform
inference on.  This argument can be a
\code{\linkS4class{DocumentTermMatrix}},
\code{\linkS4class{TermDocumentMatrix}}, a plain \code{matrix} of
document-term frequencies with column names corresponding to terms, a
\code{\linkS4class{Corpus}} or \code{\linkS4class{RSNLCorpus}} object, a
\code{\linkS4class{TokenizedDocumentView}}, or a \code{character}
object.  The default \code{\linkS4class{Tokenizer}} will be used to
tokenize objects lacking pre-specified terms, and \code{character}
objects represent single untokenized documents (if scalar) or
single tokenized documents (if vector).}
\item{var.max.iter}{The number of coordinate ascent variational
inference iterations for a single document.  A value of -1 indicates
that the process should iterate until the convergence criterion is
met.}
\item{var.convergence}{The convergence criterion for variational
inference.  Stop if (a-b)/abs(a) is less than this value, where a
is the lower bound on the likelihood for the document from the
previous variational inference step and b is the current likelihood
lower bound.}
\item{em.max.iter}{The maximum number of EM iterations to perform.}
\item{em.convergence}{The convergence criterion for EM.  Stop if
(a-b)/abs(a) is less than this value, where a is the lower bound on
the likelihood for the corpus form the previous EM step and b is the
current likelihood lower bound.}
\item{estimate.alpha}{If set to \code{TRUE} the model estimates the
Dirichlet parameter along with the topic distributions, otherwise
alpha is left fixed at its starting value.}
}


\details{
  Takes a fitted \code{\linkS4class{lda}} object and applies it to the
  document(s) represented by \code{newdata}, generating a documents by
  topics variational gamma matrix describing the extent to which each
  document in newdata is associated with each topic described by the
  fitted model.
}

\references{D. Blei, A. Ng, and M. Jordan. 2003.  Latent Dirichlet
Allocation. \emph{Journal of Machine Learning Research}, 3:993--1022.}

\author{Anthony Fader, Gary King, Dan Pemstein, and Kevin Quinn}

\seealso{
  \code{\linkS4class{lda}}
  \code{\link{lda}}
}

\keyword{methods}
