\name{TagSafe-class}
\docType{class}
\alias{TagSafe-class}

\title{Class "TagSafe"}

\description{A virtual type that other classes may extend to indicate
that they correctly handle \code{\linkS4class{Tagged}} vectors.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\author{Anthony Fader, Gary King, Dan Pemstein, and Kevin Quinn}

\seealso{
  \code{\linkS4class{Tagged}}
}

\keyword{classes}
