\name{TokenTransform-class}
\docType{class}
\alias{TokenTransform-class}
\alias{TokenTransform}
\alias{transformTokens,character,TokenTransform,missing,missing-method}
\alias{transformTokens,character,TokenTransform,Tokenizer,missing-method}
\alias{transformTokens,TokenizedCorpusView,TokenTransform,missing,missing-method}
\alias{transformTokens,TokenizedCorpusView,TokenTransform,missing,numeric-method}
\alias{transformTokens,RSNLCorpus,TokenTransform,ANY,missing-method}
\alias{transformTokens,RSNLCorpus,TokenTransform,ANY,numeric-method}
\alias{transformTokens,TokenizedDocumentView,TokenTransform,missing,missing-method}

\title{Class "TokenTransform"}
\description{
  An abstract class representing an object that takes a sequence of
  tokens  and returns a modified version of the original.
}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
  \describe{
    \item{transformTokens}{\code{signature(object = "character",
                                        transform = "TokenTransform",
                                        tokenizer = "missing",
                                        index = "missing")}:
      Apply the given transform to the \code{character} object and return
      a \code{character} vector of modified tokens.}

    \item{transformTokens}{\code{signature(object = "character",
                                        transform = "TokenTransform",
                                        tokenizer = "Tokenizer",
                                        index = "missing")}:
      Apply the given transform to the \code{character} object, after first
      tokenizing with the given \code{\linkS4class{Tokenizer}}, and
      return a \code{character} vector of modified tokens. }

    \item{transformTokens}{\code{signature(object = "TokenizedCorpusView",
                                        transform = "TokenTransform",
                                        tokenizer = "missing",
                                        index = "missing")}:
      Apply the given transform to the
      \code{\linkS4class{TokenizedCorpusView}} and return a modified
      view.}

    \item{transformTokens}{\code{signature(object = "TokenizedCorpusView",
                                        transform = "TokenTransform",
                                        tokenizer = "missing",
                                        index = "numeric")}:
      Apply the transform to the
      \code{\linkS4class{TokenizedDocumentView}} at \code{index}
      within the \code{\linkS4class{TokenizedCorpusView}} and return a
      modified view.}

    \item{transformTokens}{\code{signature(object = "RSNLCorpus",
                                        transform = "TokenTransform",
                                        tokenizer = "ANY",
                                        index = "missing")}:
      Apply the transform to the
      \code{\linkS4class{TokenizedCorpusView}} generated by applying
      the given \code{\linkS4class{Tokenizer}} (or the default
      tokenizer, if none is given) to \code{object} and return a
      \code{\linkS4class{TokenizedCorpusView}}.}
      
    \item{transformTokens}{\code{signature(object = "RSNLCorpus",
                                        transform = "TokenTransform",
                                        tokenizer = "ANY",
                                         index = "numeric")}:
      Apply the transform to the
      \code{\linkS4class{TokenizedDocumentView}} generated by applying
      the given \code{\linkS4class{Tokenizer}} (or the default
      tokenizer, if none is given) to the
      \code{\linkS4class{TextDocument}} at \code{index} within
      \code{object} and return a
      \code{\linkS4class{TokenizedDocumentView}}.}

    \item{transformTokens}{\code{signature(object = "TokenizedDocumentView",
                                        transform = "TokenTransform",
                                        tokenizer = "missing",
                                        index = "missing")}:
      Apply the given transform to the
      \code{\linkS4class{TokenizedDocumentView}} and return a modified
      view.}

	 }
}


\author{Anthony Fader, Gary King, Dan Pemstein, and Kevin Quinn}

\seealso{
  \code{\link{transformTokens}}
  \code{\linkS4class{FunctionalTokenTransform}}
  \code{\linkS4class{RegexTokenTransform}}
  \code{\linkS4class{Stemmer}}
  \code{\linkS4class{TokenizedView}}
  \code{\linkS4class{TokenizedDocumentView}}
  \code{\linkS4class{TokenizedCorpusView}}
}



\keyword{classes}
