\name{TolowerTokenTransform-class}
\docType{class}
\alias{TolowerTokenTransform-class}

\title{Class "TolowerTokenTransform"}

\description{A \code{\linkS4class{FunctionalTokenTransform}} for
converting tokens to lower case.}

\section{Objects from the Class}{
Objects can be created by calls of the form
\code{TolowerTokenTransform()}
}

\section{Extends}{
Class \code{"\linkS4class{TagSafeFunctionalTokenTransform}"}, directly.
Class \code{"\linkS4class{FunctionalTokenTransform}"}, by class "TagSafeFunctionalTokenTransform", distance 2.
Class \code{"\linkS4class{TagSafe}"}, by class "TagSafeFunctionalTokenTransform", distance 2.
Class \code{"\linkS4class{TokenTransform}"}, by class "TagSafeFunctionalTokenTransform", distance 3.
}

\author{Anthony Fader, Gary King, Dan Pemstein, and Kevin Quinn}

\seealso{
  \code{\link{transformTokens}}
  \code{\linkS4class{FunctionalTokenTransform}}
  \code{\linkS4class{TokenTransform}}
}

\keyword{classes}
